Date and formatting updates
I found that if I updated column F to 0.00 format as instructed, it also updated the header. Obviously it should be 2015, not 2,015.00 but if we're changing the name of the column anyway we might as well apply the format to the entirety of column F and avoid this little "gotcha".